refactor: use only soundfile backend
